+++ tests/Bridge/Symfony/Normalizer/Data.php
@@ -4,17 +4,11 @@ declare(strict_types=1);
 
 namespace steevanb\PhpTypedArray\Tests\Bridge\Symfony\Normalizer;
 
-final class TestObject
+class Data
 {
     /** @var string */
     public $foo;
 
     /** @var int */
     public $bar;
-
-    public function __construct(string $foo, int $bar)
-    {
-        $this->foo = $foo;
-        $this->bar = $bar;
-    }
 }