+++ tests/Bridge/Symfony/Normalizer/StringArrayDenormalizerTest.php
@@ -7,20 +7,20 @@ namespace steevanb\PhpTypedArray\Tests\Bridge\Symfony\Normalizer;
 use PHPUnit\Framework\TestCase;
 use steevanb\PhpTypedArray\{
     Bridge\Symfony\Normalizer\ScalarArrayDenormalizer,
-    ScalarArray\FloatArray
+    ScalarArray\StringArray
 };
 use Symfony\Component\Serializer\Serializer;
 
-final class FloatArrayNormalizerTest extends TestCase
+final class StringArrayDenormalizerTest extends TestCase
 {
     public function testDenormalize(): void
     {
         $serializer = new Serializer([new ScalarArrayDenormalizer()]);
-        /** @var FloatArray $array */
-        $array = $serializer->denormalize([1.0, 2.2], FloatArray::class);
+        /** @var StringArray $array */
+        $array = $serializer->denormalize(['string1', 'string2'], StringArray::class);
 
         static::assertCount(2, $array);
-        static::assertSame(1.0, $array[0]);
-        static::assertSame(2.2, $array[1]);
+        static::assertSame('string1', $array[0]);
+        static::assertSame('string2', $array[1]);
     }
 }